#17_reworked_user_editor_and_added_campaings_list

+++ src/main/java/com/test/entities/Campaign.java
@@ -38,7 +38,7 @@ public class Campaign {
     private Date dateOfExpiry;
     private String logo;
 
-    public Integer getAuthorId() {
+    public Long getAuthorId() {
         return author.getId();
     }
