+++ src/main/java/com/test/entities/Payment.java
@@ -24,7 +24,7 @@ public class Payment {
     private Double amount;
     private Date dateOfCreation;
 
-    public Integer getAuthorId() {
+    public Long getAuthorId() {
         return author.getId();
     }
